pokemmo-wrapped

+++ pkgs/pokemmo-wrapped/default.nix
@@ -5,9 +5,12 @@
 , wget
 , unzip
 , writeText
+, makeDesktopItem
+, fetchzip
+, copyDesktopItems
 }:
 
-stdenv.mkDerivation {
+stdenv.mkDerivation rec {
   pname = "pokemmo-wrapped";
   version = "1.0.0";
 
@@ -16,9 +19,25 @@ stdenv.mkDerivation {
     makeWrapper
     wget
     unzip
+    copyDesktopItems
   ];
 
-  dontUnpack = true;
+  src = fetchzip {
+    url = "https://dl.pokemmo.com/PokeMMO-Client.zip";
+    sha256 = "sha256-OkRpDPOgFKKpZO/8K7qA3WiLN7nbV0UbV+EMUBMWA70";
+    stripRoot=false;
+  };
+
+  desktopItems = [
+    (makeDesktopItem {
+      name = pname;
+      desktopName = "PokeMMO";
+      exec = "pokemmo";
+      comment = meta.description;
+      icon = "pokemmo";
+      categories = [ "Game" ];
+    })
+  ];
 
   pokemmoWrappedText = writeText "pokemmo-wrapped" ''
 # Unninstall
@@ -27,72 +46,51 @@ do
   if [ "$c" == "-u" ]; 
   then
   rm -r ~/.local/share/pokemmo/ 
-  rm ~/.local/share/icons/hicolor/128x128/apps/pokemmo.png 
-  rm ~/.local/share/applications/pokemmo.desktop
-  rm ~/.local/bin/pokemmo
+
   exit 0
   fi
 done
 
-# Fetching
-wget -O pokemmo.zip https://dl.pokemmo.com/PokeMMO-Client.zip
+if [ ! -e "$HOME/.local/share/pokemmo/" ]; 
+  then
+  echo "As you dont have pokemmo installed, now the installation will start!"
 
-# Moving to local
-mkdir -p ~/.local/share/pokemmo/
-cp -r pokemmo.zip ~/.local/share/pokemmo/
-rm pokemmo.zip
-cd ~/.local/share/pokemmo/
-unzip -qq pokemmo.zip
-rm pokemmo.zip
+  # Fetching
+  wget -q -O pokemmo.zip https://dl.pokemmo.com/PokeMMO-Client.zip
 
-# Icons
-mkdir -p ~/.local/share/icons/hicolor/128x128/apps
-cp -r ~/.local/share/pokemmo/data/icons/128x128.png ~/.local/share/icons/hicolor/128x128/apps/pokemmo.png
+  # Moving to local
+  mkdir -p ~/.local/share/pokemmo/
+  cp -r pokemmo.zip ~/.local/share/pokemmo/
+  rm pokemmo.zip
+  cd ~/.local/share/pokemmo/
+  unzip -qq pokemmo.zip
+  rm pokemmo.zip
 
-# Desktop entry
-touch pokemmo.desktop
-echo "[Desktop Entry]
-Name=PokeMMO
-GenericName=Text Editor
-Comment=MMO based on the Pokémon Universe
-TryExec=pokemmo
-Exec=pokemmo
-Terminal=false
-Type=Application
-Keywords=game;pokemon;
-Icon=pokemmo
-Categories=Games;
-StartupNotify=false
-" >> pokemmo.desktop 
-mkdir -p ~/.local/share/applications/ 
-cp -r pokemmo.desktop ~/.local/share/applications/
-rm pokemmo.desktop
+  echo "PokeMMO has been Installed!" 
+fi
 
-# Execution Script
-touch pokemmo
-
-echo "export PATH="${jdk17.out}/bin:\$PATH"
-cd ~/.local/share/pokemmo/
-exec ./PokeMMO.sh" >> pokemmo
-
-chmod +x pokemmo
-mkdir -p ~/.local/bin/
-cp -r pokemmo ~/.local/bin/
-rm pokemmo
-
-echo "PokeMMO has been Installed!" 
+cd ~/.local/share/pokemmo
+./PokeMMO.sh
   '';
 
 
   buildPhase = ''
     cp $pokemmoWrappedText pokemmo-wrapped
     chmod +x pokemmo-wrapped
+    cp data/icons/128x128.png .
     '';
 
   installPhase = ''
     mkdir -p $out/bin
     mkdir -p $out/share
     
+    #Copy the desktop item into $out
+    copyDesktopItems
+
+    # Copy icon files
+    mkdir -p $out/share/icons/hicolor/128x128/apps
+    cp 128x128.png $out/share/icons/hicolor/128x128/apps/pokemmo.png 
+
     cp pokemmo-wrapped $out/share/pokemmo-wrapped
     ls $out/share/
     ln -s $out/share/pokemmo-wrapped $out/bin/pokemmo
